+++ Sources/KeyboardKit/Feedback/AudioFeedback.swift
@@ -12,8 +12,11 @@ import Foundation
  This enum contains audio feedback types that maps to system
  audio feedback values.
  
- You can call ``trigger()`` on any feedback value to trigger
- it using ``AudioFeedbackEngine/shared``.
+ Each feedback type has a unique system sound ``id`` that is
+ used when the sound is played with ``AudioFeedback/Engine``.
+ 
+ You can call ``trigger()`` on any feedback type, to play it
+ with the ``AudioFeedback/Engine/shared`` audio engine.
 */
 public enum AudioFeedback: Codable, Equatable, Identifiable {
     
@@ -37,12 +40,7 @@ public enum AudioFeedback: Codable, Equatable, Identifiable {
 
 public extension AudioFeedback {
     
-    /**
-     The unique feedback identifier.
-
-     This identifier maps to a unique system sound, which is
-     used by the ``AudioFeedbackEngine``.
-     */
+    /// The unique feedback identifier.
     var id: UInt32 {
         switch self {
         case .input: return 1104
@@ -55,6 +53,6 @@ public extension AudioFeedback {
     
     /// Trigger the feedback with the shared feedback engine.
     func trigger() {
-        AudioFeedbackEngine.shared.trigger(self)
+        AudioFeedback.Engine.shared.trigger(self)
     }
 }